Move DoctrineCollectionAdapter to package (with B/C)
