+++ lib/Adapter/DoctrineCollections/DoctrineCollectionsAdapter.php
@@ -1,0 +1,51 @@
+<?php
+
+namespace Pagerfanta\DoctrineCollections;
+
+use Doctrine\Common\Collections\Collection;
+use Pagerfanta\Adapter\AdapterInterface;
+
+/**
+ * Adapter which calculates pagination from a Doctrine Collection.
+ */
+class DoctrineCollectionsAdapter implements AdapterInterface
+{
+    /**
+     * @var Collection
+     */
+    private $collection;
+
+    public function __construct(Collection $collection)
+    {
+        $this->collection = $collection;
+    }
+
+    /**
+     * Retrieves the Collection.
+     *
+     * @return Collection
+     */
+    public function getCollection()
+    {
+        return $this->collection;
+    }
+
+    /**
+     * @return int
+     */
+    public function getNbResults()
+    {
+        return $this->collection->count();
+    }
+
+    /**
+     * @param int $offset
+     * @param int $length
+     *
+     * @return iterable
+     */
+    public function getSlice($offset, $length)
+    {
+        return $this->collection->slice($offset, $length);
+    }
+}
